Tweak Blender exporter to handle .mtls with options
Currently, Blender doesn't parse .mtl texture lines with options _and_ a
file path correctly -- the options turn up in the 'file_path' value of
the texture entry. The Three.js exporter can't handle these when copying
textures, because they're not valid file paths.

This tweaks the bundled exporter to try and extract the right file path,
assuming that it's the last entry in the .mtl line, and use that for the
copy command.